chore: refactor to use modules
