--- data.txt
@@ -1,4 +1,0 @@
-50
-12
-13
-12
