+++ go.mod
@@ -1,0 +1,3 @@
+module github.com/sudo-kaizen/go-mmm
+
+go 1.15
